requirements.txt
asgiref==3.3.4
astroid==2.5.6
Django==3.2.3
djangorestframework==3.12.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
